+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
+- Require TYPO3 7.6 and drop support for TYPO3 6.2 (#67)
 - Drop support for PHP 5.5 (#64)
 
 ### Fixed